// src/app/models/projectdetails.ts
export interface Project {
    project_id: string;              
    project_name: string;            
    project_head: string;             
    start_date: string;               
    end_date: string;
    created_on: string;
    created_by: string; 
    modified_on: string;
    modified_by: string;
}